CrowdsaleEndTime: update test
